Fixed warning. Moved kate mods to files
