+++ kate/recovery/src/proof.rs
@@ -1,3 +1,9 @@
+use thiserror_no_std::Error;
+
+#[cfg(feature = "std")]
+use crate::{data::Cell, matrix::Dimensions};
+#[cfg(feature = "std")]
+use avail_core::constants::kate::COMMITMENT_SIZE;
 #[cfg(feature = "std")]
 use dusk_bytes::Serializable;
 #[cfg(feature = "std")]
@@ -7,20 +13,18 @@ use dusk_plonk::{
 	fft::EvaluationDomain,
 	prelude::BlsScalar,
 };
-use thiserror_no_std::Error;
-
-use crate::{data::Cell, matrix::Dimensions};
-use avail_core::constants::kate::COMMITMENT_SIZE;
-
+#[cfg(feature = "std")]
+use poly_multiproof::traits::AsBytes;
+#[cfg(feature = "std")]
 use poly_multiproof::{
 	ark_poly::{EvaluationDomain as ArkEvaluationDomain, GeneralEvaluationDomain},
 	m1_blst::{Bls12_381, Fr, M1NoPrecomp, Proof as ArkProof},
 	traits::KZGProof,
 };
-
+#[cfg(feature = "std")]
 type ArkScalar = poly_multiproof::m1_blst::Fr;
+#[cfg(feature = "std")]
 type ArkCommitment = poly_multiproof::Commitment<Bls12_381>;
-use poly_multiproof::traits::AsBytes;
 
 #[derive(Error, Debug)]
 pub enum Error {
